Remove the linear regression line on the plot
Students are unfamiliar with the concept of linear regression at the time when they are working with this notebook. This part of the plot causes confusion (e.g. is the line part of the data points? what does the line mean?) and side discussions from students who already know about linear regression (e.g. what's the error with the linear regression in Seaborn?). This part of the plot should be also be removed because it is not pertinent to the content of the notebook.