Update 3_keras_basic_feat_eng.ipynb
Line 88, 89: I edited this line for grammar and capping, but maybe I shouldn't have? 
Line 1153: I'm curious about the use of "again." Did the lab already make this point, or is this a repeated step? 
Google style is only one space between sentences, but that might be difficult to implement using find/replace.