rewrite strategy definition
switching the strategy definition and moving from TF2.1 to TF2.3.2 reduces the run time of the TPU command from 34 minutes to 14 minutes wall time.
we also switch from TPUv2 to TPUv3.
we need to include distribute_utils.py from the official tensorflow package (this could also be installed at tf-models-official but that would introduce an install dependency).
in the `gcloud compute ... create` command, the machine-type is explicitly set to be safe. The other changes are described above.
The TF runner requires the TPU_NAME environmental variable to be set, hence the new line before the `python3 -m` command.