Wrong indentation is making get_scal return None
The function get_scal should return the resulting minmax function, but, due to the wrong indentation in the outer return statement, it is returning None. This does not provoke an error since, further down, the argument normalizer_fn of numeric_column accepts None, but the result of training is wrong (as if no normalization was applied).