Fix the "input_dim" of the dense layer
After adding more features, including bucketized feature and feature crosses, the number of features is no longer 8. The number of 8 comes from the first model, which only has numeric features. This is a silent error and model trains without error but the whole goal of this tutorial is to add more features and reduce the "mean squared error". After changing this line, my `mse` reduced from `14850733056` to `14575827968`.